fixed dct

+++ pipeline/stages/decorrelation/strategy/transformation_strategy.py
@@ -1,6 +1,7 @@
 import numpy as np
 from numba import njit
 
+
 @njit
 def precompute_cosines(N):
     cosines = np.zeros((N, N), dtype=np.float64)
@@ -8,6 +9,7 @@ def precompute_cosines(N):
         for n in range(N):
             cosines[k, n] = np.cos((2 * n + 1) * k * np.pi / (2 * N))
     return cosines
+
 
 @njit
 def dct_block(block, cosines, block_size):
@@ -23,6 +25,7 @@ def dct_block(block, cosines, block_size):
             result[u, v] = 0.25 * Cu * Cv * sum_val
     return result
 
+
 @njit
 def idct_block(block, cosines, block_size):
     result = np.zeros((block_size, block_size), dtype=np.float64)
@@ -37,6 +40,7 @@ def idct_block(block, cosines, block_size):
             result[x, y] = 0.25 * sum_val
     return result
 
+
 @njit
 def process_blocks(frame, height, width, block_size, cosines, inverse=False):
     padded_h = ((height + block_size - 1) // block_size) * block_size
@@ -47,12 +51,17 @@ def process_blocks(frame, height, width, block_size, cosines, inverse=False):
     result = np.zeros_like(padded)
     for i in range(0, padded_h, block_size):
         for j in range(0, padded_w, block_size):
-            block = padded[i:i + block_size, j:j + block_size]
+            block = padded[i : i + block_size, j : j + block_size]
             if inverse:
-                result[i:i + block_size, j:j + block_size] = idct_block(block, cosines, block_size)
+                result[i : i + block_size, j : j + block_size] = idct_block(
+                    block, cosines, block_size
+                )
             else:
-                result[i:i + block_size, j:j + block_size] = dct_block(block, cosines, block_size)
+                result[i : i + block_size, j : j + block_size] = dct_block(
+                    block, cosines, block_size
+                )
     return result[:height, :width]
+
 
 class BlockDCT:
     def __init__(self, width, height, block_size=8):
@@ -62,13 +71,22 @@ class BlockDCT:
         self.cosines = precompute_cosines(block_size)
 
     def transform(self, frame: np.ndarray) -> np.ndarray:
-        return process_blocks(frame, self.height, self.width, self.block_size, self.cosines, inverse=False)
+        return process_blocks(
+            frame, self.height, self.width, self.block_size, self.cosines, inverse=False
+        )
 
     def inverse_transform(self, transformed: np.ndarray) -> np.ndarray:
-        result = process_blocks(transformed, self.height, self.width, self.block_size, self.cosines, inverse=True)
-        #tengo un margen de error en la transformación inversa
-        #todos los pixeles tiene un valor -1 de lo que deberían ser
-        #esto es porque la DCT no es exacta, y la IDCT tampoco
-        #por lo tanto, redondeo el resultado para que sea un entero
-        #y lo devuelvo como uint8
-        return np.round(result)  
+        result = process_blocks(
+            transformed,
+            self.height,
+            self.width,
+            self.block_size,
+            self.cosines,
+            inverse=True,
+        )
+        # tengo un margen de error en la transformación inversa
+        # todos los pixeles tiene un valor -1 de lo que deberían ser
+        # esto es porque la DCT no es exacta, y la IDCT tampoco
+        # por lo tanto, redondeo el resultado para que sea un entero
+        # y lo devuelvo como uint8
+        return np.round(result).astype(np.uint8)
